+++ lib/main.dart
@@ -3,9 +3,8 @@ import 'package:flutter_svg/svg.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:kidoresources/constants.dart';
 import 'package:kidoresources/screens/CalenderScreen.dart';
-import 'package:kidoresources/screens/widgets/Category_Card.dart';
-import 'package:kidoresources/screens/widgets/Navigation_bar.dart';
-import 'package:kidoresources/screens/widgets/search_bar.dart';
+import 'package:kidoresources/screens/HomeScreen.dart';
+
 
 void main() => runApp(MyApp());
 
@@ -25,98 +24,5 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class HomeScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    var size = MediaQuery.of(context).size;
-    return Scaffold(
-      bottomNavigationBar: Container(
-        padding: EdgeInsets.symmetric(horizontal: 70, vertical: 10),
-        height: 70,
-        color: Colors.grey[200],
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: <Widget>[
-            BottomNavigation(
-              title: "Today",
-              svgSrc: "assets/icons/Icon_calendar.svg",
-              press: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) {
-                    return CalenderScreen();
-                  }),
-                );
-              },
-              isActive: false,
-            ),
-            BottomNavigation(
-              title: "Home",
-              svgSrc: "assets/icons/Icon_home.svg",
-              press: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) {
-                    return HomeScreen();
-                  }),
-                );
-              },
-              isActive: true,
-            ),
-          ],
-        ),
-      ),
-      body: Stack(
-        children: <Widget>[
-          Container(
-            height: size.height * .45,
-            decoration: BoxDecoration(
-              color: Color(0xFFF5CEB8),
-              image: DecorationImage(
-                  alignment: Alignment.centerLeft,
-                  image: AssetImage("assets/images/undraw_pilates_gpdb.png")),
-            ),
-          ),
-          SafeArea(
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Text(
-                    "Good Morning \nPhantayisa",
-                    style: TextStyle(
-                      fontWeight: FontWeight.w900,
-                      fontSize: 35,
-                    ),
-                  ),
-                  SearchBar(),
-                  Expanded(
-                    child: GridView.count(
-                      crossAxisCount: 2,
-                      childAspectRatio: .85,
-                      crossAxisSpacing: 10,
-                      mainAxisSpacing: 10,
-                      children: <Widget>[
-                        CatagoryCard(
-                          title: "Food Resources",
-                          svgSrc: "assets/icons/Hamburger.svg",
-                          press: () {},
-                        ),
-                        CatagoryCard(
-                          title: "Actvities",
-                          svgSrc: "assets/icons/yoga.svg",
-                          press: () {},
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
+
+
